Add start of API Login process
Right now, the backend is capable of querying the comments from the LW website and forward them to the frontend. As next step a DB has to be added to actually use real data, instead of dummy data.

Error handling is still quite hacky. Correct practices and error codes probably not used. But it works so far.