Refactor backend: Clean up all SQL code & unique handling
Previously multiple insert statements had been used in case of collision, this needlessly increases the id counter in the DB.
It is better to create and entry and update it. In the worst case this wastes one index + it allows for incomplete data though.

The maximum attempts for unique injection has been reduced to 10 times, instead of unlimited for half a second. As chances of collision are very low anyway.

The unique injection code is now inside its own class.

Several other accesses to PDO have been improved or helper methods have been added to reduce code size and make it cleaner.
While catch blocks delete calls are now silent (errors are discarded there).